Cache SslClientAuthenticationOptions on SocketsHttpHandler pool
For the common case, we can just use a singleton cached instance with settings computed when the pool is created.  Only if it's accessed via HttpClientHandler with a server certificate callback do we need to still create an instance per request, in order to wrap the delegate to provide the expected semantics.